# docs/BeneficiaryAPI.Swagger20.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Beneficiary API",
    "contact": {}
  },
  "host": "app:3000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/beneficiary": {
      "post": {
        "summary": "Create beneficiary",
        "tags": ["beneficiary"],
        "operationId": "Createbeneficiary",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreatebeneficiaryRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Createbeneficiary"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "id": "ece42f82-48d6-4eda-b6f6-ed2be7186189"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "45"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"2d-GGmU6l5RP/YWQCikGcRjdt7mmE4\""
              },
              "Date": {
                "type": "string",
                "default": "Tue, 15 Jun 2021 01:30:43 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
    "/beneficiaries": {
      "get": {
        "summary": "Get beneficiaries",
        "tags": ["beneficiary"],
        "operationId": "Getbeneficiaries",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "searchValue",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Getbeneficiary"
              }
            },
            "examples": {
              "application/json; charset=utf-8": [
                {
                  "id": "ece42f82-48d6-4eda-b6f6-ed2be7186189",
                  "name": "Ronaldo Silva Pereira",
                  "email": "ronaldo.pereira@gmail.com",
                  "document": "53902371021",
                  "documentType": "CPF",
                  "status": "DRAFT",
                  "agencyNumber": "0001",
                  "agencyDigit": null,
                  "accountNumber": "123456",
                  "accountDigit": "1",
                  "createdAt": "2021-06-15T01:30:43Z",
                  "updatedAt": "2021-06-15T01:30:43Z",
                  "bankId": "BANCO_DO_BRASIL",
                  "bankAccounTypeId": 1,
                  "Bank": {
                    "id": "BANCO_DO_BRASIL",
                    "name": "Banco do Brasil (001)",
                    "createdAt": "2021-06-15T01:19:34Z",
                    "updatedAt": "2021-06-15T01:19:34Z"
                  },
                  "BankAccountType": {
                    "id": 1,
                    "type": "CONTA_CORRENTE",
                    "name": "Conta corrente",
                    "createdAt": "2021-06-15T01:19:50Z",
                    "updatedAt": "2021-06-15T01:19:50Z",
                    "bankId": "BANCO_DO_BRASIL"
                  }
                }
              ]
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "709"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"2c5-oZZD6yFFu75TID0oeUYwuexOPv0\""
              },
              "Date": {
                "type": "string",
                "default": "Tue, 15 Jun 2021 03:26:17 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Batch delete beneficiaries",
        "tags": ["beneficiary"],
        "operationId": "Batchdeletebeneficiaries",
        "deprecated": false,
        "produces": ["text/plain; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/BatchdeletebeneficiariesRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            },
            "examples": {
              "text/plain; charset=utf-8": "OK"
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "2"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"2-nOO9QiTIwXgNtWtBJezz8kv3SLc\""
              },
              "Date": {
                "type": "string",
                "default": "Tue, 15 Jun 2021 22:41:32 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
    "/beneficiary/{id}": {
      "get": {
        "summary": "Get beneficiary",
        "tags": ["beneficiary"],
        "operationId": "Getbeneficiary",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Getbeneficiary1"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "id": "350528aa-f97a-4acd-8e57-d9e2abd16951",
                "name": "Ronaldo Silva Pereira",
                "email": "ronaldo.pereira@gmail.com",
                "document": "53902371021",
                "documentType": "CPF",
                "status": "DRAFT",
                "agencyNumber": "0001",
                "agencyDigit": null,
                "accountNumber": "123456",
                "accountDigit": "1",
                "createdAt": "2021-06-15T23:09:03Z",
                "updatedAt": "2021-06-15T23:09:03Z",
                "bankId": "BANCO_DO_BRASIL",
                "bankAccounTypeId": 1
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "386"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"182-y3J+p4fQPTmjLYT0TpAShmTGadk\""
              },
              "Date": {
                "type": "string",
                "default": "Tue, 15 Jun 2021 23:09:13 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update beneficiary",
        "tags": ["beneficiary"],
        "operationId": "Updatebeneficiary",
        "deprecated": false,
        "produces": ["text/plain; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdatebeneficiaryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            },
            "examples": {
              "text/plain; charset=utf-8": "OK"
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "2"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"2-nOO9QiTIwXgNtWtBJezz8kv3SLc\""
              },
              "Date": {
                "type": "string",
                "default": "Tue, 15 Jun 2021 22:24:50 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "CreatebeneficiaryRequest": {
      "title": "CreatebeneficiaryRequest",
      "example": {
        "name": "Ronaldo Silva Pereira",
        "email": "ronaldo.pereira@gmail.com",
        "document": "53902371021",
        "documentType": "CPF",
        "agencyNumber": "0001",
        "agencyDigit": null,
        "accountNumber": "123456",
        "accountDigit": "1",
        "bankId": "BANCO_DO_BRASIL",
        "bankAccounTypeId": "1"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "document": {
          "type": "string"
        },
        "documentType": {
          "type": "string"
        },
        "agencyNumber": {
          "type": "string"
        },
        "agencyDigit": {
          "type": "string"
        },
        "accountNumber": {
          "type": "string"
        },
        "accountDigit": {
          "type": "string"
        },
        "bankId": {
          "type": "string"
        },
        "bankAccounTypeId": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "email",
        "document",
        "documentType",
        "agencyNumber",
        "agencyDigit",
        "accountNumber",
        "accountDigit",
        "bankId",
        "bankAccounTypeId"
      ]
    },
    "Createbeneficiary": {
      "title": "Createbeneficiary",
      "example": {
        "id": "ece42f82-48d6-4eda-b6f6-ed2be7186189"
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      },
      "required": ["id"]
    },
    "Getbeneficiary": {
      "title": "Getbeneficiary",
      "example": {
        "id": "ece42f82-48d6-4eda-b6f6-ed2be7186189",
        "name": "Ronaldo Silva Pereira",
        "email": "ronaldo.pereira@gmail.com",
        "document": "53902371021",
        "documentType": "CPF",
        "status": "DRAFT",
        "agencyNumber": "0001",
        "agencyDigit": null,
        "accountNumber": "123456",
        "accountDigit": "1",
        "createdAt": "2021-06-15T01:30:43Z",
        "updatedAt": "2021-06-15T01:30:43Z",
        "bankId": "BANCO_DO_BRASIL",
        "bankAccounTypeId": 1,
        "Bank": {
          "id": "BANCO_DO_BRASIL",
          "name": "Banco do Brasil (001)",
          "createdAt": "2021-06-15T01:19:34Z",
          "updatedAt": "2021-06-15T01:19:34Z"
        },
        "BankAccountType": {
          "id": 1,
          "type": "CONTA_CORRENTE",
          "name": "Conta corrente",
          "createdAt": "2021-06-15T01:19:50Z",
          "updatedAt": "2021-06-15T01:19:50Z",
          "bankId": "BANCO_DO_BRASIL"
        }
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "document": {
          "type": "string"
        },
        "documentType": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "agencyNumber": {
          "type": "string"
        },
        "agencyDigit": {
          "type": "string"
        },
        "accountNumber": {
          "type": "string"
        },
        "accountDigit": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "bankId": {
          "type": "string"
        },
        "bankAccounTypeId": {
          "type": "integer",
          "format": "int32"
        },
        "Bank": {
          "$ref": "#/definitions/Bank1"
        },
        "BankAccountType": {
          "$ref": "#/definitions/BankAccountType"
        }
      },
      "required": [
        "id",
        "name",
        "email",
        "document",
        "documentType",
        "status",
        "agencyNumber",
        "agencyDigit",
        "accountNumber",
        "accountDigit",
        "createdAt",
        "updatedAt",
        "bankId",
        "bankAccounTypeId",
        "Bank",
        "BankAccountType"
      ]
    },
    "Bank1": {
      "title": "Bank1",
      "example": {
        "id": "BANCO_DO_BRASIL",
        "name": "Banco do Brasil (001)",
        "createdAt": "2021-06-15T01:19:34Z",
        "updatedAt": "2021-06-15T01:19:34Z"
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      },
      "required": ["id", "name", "createdAt", "updatedAt"]
    },
    "BankAccountType": {
      "title": "BankAccountType",
      "example": {
        "id": 1,
        "type": "CONTA_CORRENTE",
        "name": "Conta corrente",
        "createdAt": "2021-06-15T01:19:50Z",
        "updatedAt": "2021-06-15T01:19:50Z",
        "bankId": "BANCO_DO_BRASIL"
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "bankId": {
          "type": "string"
        }
      },
      "required": ["id", "type", "name", "createdAt", "updatedAt", "bankId"]
    },
    "Getbeneficiary1": {
      "title": "Getbeneficiary1",
      "example": {
        "id": "350528aa-f97a-4acd-8e57-d9e2abd16951",
        "name": "Ronaldo Silva Pereira",
        "email": "ronaldo.pereira@gmail.com",
        "document": "53902371021",
        "documentType": "CPF",
        "status": "DRAFT",
        "agencyNumber": "0001",
        "agencyDigit": null,
        "accountNumber": "123456",
        "accountDigit": "1",
        "createdAt": "2021-06-15T23:09:03Z",
        "updatedAt": "2021-06-15T23:09:03Z",
        "bankId": "BANCO_DO_BRASIL",
        "bankAccounTypeId": 1
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "document": {
          "type": "string"
        },
        "documentType": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "agencyNumber": {
          "type": "string"
        },
        "agencyDigit": {
          "type": "string"
        },
        "accountNumber": {
          "type": "string"
        },
        "accountDigit": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "bankId": {
          "type": "string"
        },
        "bankAccounTypeId": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "id",
        "name",
        "email",
        "document",
        "documentType",
        "status",
        "agencyNumber",
        "agencyDigit",
        "accountNumber",
        "accountDigit",
        "createdAt",
        "updatedAt",
        "bankId",
        "bankAccounTypeId"
      ]
    },
    "UpdatebeneficiaryRequest": {
      "title": "UpdatebeneficiaryRequest",
      "example": {
        "name": "Rolandinho Fernandes",
        "email": "rolandinho@gmail.com",
        "documentType": "CNPJ",
        "document": "88960753000190",
        "agencyNumber": "0009",
        "agencyDigit": null,
        "accountNumber": "098765",
        "accountDigit": "2",
        "bankId": null,
        "bankAccountTypeId": null
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "documentType": {
          "type": "string"
        },
        "document": {
          "type": "string"
        },
        "agencyNumber": {
          "type": "string"
        },
        "agencyDigit": {
          "type": "string"
        },
        "accountNumber": {
          "type": "string"
        },
        "accountDigit": {
          "type": "string"
        },
        "bankId": {
          "type": "string"
        },
        "bankAccountTypeId": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "email",
        "documentType",
        "document",
        "agencyNumber",
        "agencyDigit",
        "accountNumber",
        "accountDigit",
        "bankId",
        "bankAccountTypeId"
      ]
    },
    "BatchdeletebeneficiariesRequest": {
      "title": "BatchdeletebeneficiariesRequest",
      "example": {
        "ids": ["ece42f82-48d6-4eda-b6f6-ed2be7186189"]
      },
      "type": "object",
      "properties": {
        "ids": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": ["ids"]
    }
  },
...
}